app/requirements.txt
bs4
fastapi
requests
termcolor
python-dotenv
uvicorn